+++ llama_index/agent/react/base.py
@@ -1,5 +1,3 @@
-# ReAct agent
-
 import asyncio
 from threading import Thread
 from typing import Any, Dict, List, Optional, Sequence, Tuple, Type, cast
@@ -25,7 +23,7 @@ from llama_index.llms.openai import OpenAI
 from llama_index.memory.chat_memory_buffer import ChatMemoryBuffer
 from llama_index.memory.types import BaseMemory
 from llama_index.objects.base import ObjectRetriever
-from llama_index.tools import BaseTool, adapt_to_async_tool
+from llama_index.tools import BaseTool, ToolOutput, adapt_to_async_tool
 from llama_index.tools.types import AsyncBaseTool
 from llama_index.utils import print_text
 
@@ -39,7 +37,6 @@ class ReActAgent(BaseAgent):
     completion endpoints.
 
     Can take in a set of tools that require structured inputs.
-
     """
 
     def __init__(
@@ -54,13 +51,14 @@ class ReActAgent(BaseAgent):
         verbose: bool = False,
         tool_retriever: Optional[ObjectRetriever[BaseTool]] = None,
     ) -> None:
+        super().__init__(callback_manager=callback_manager or llm.callback_manager)
         self._llm = llm
         self._memory = memory
         self._max_iterations = max_iterations
         self._react_chat_formatter = react_chat_formatter or ReActChatFormatter()
         self._output_parser = output_parser or ReActOutputParser()
-        self.callback_manager = callback_manager or self._llm.callback_manager
         self._verbose = verbose
+        self.sources: List[ToolOutput] = []
 
         if len(tools) > 0 and tool_retriever is not None:
             raise ValueError("Cannot specify both tools and tool_retriever")
@@ -88,16 +86,14 @@ class ReActAgent(BaseAgent):
         verbose: bool = False,
         **kwargs: Any,
     ) -> "ReActAgent":
-        tools = tools or []
-        chat_history = chat_history or []
         llm = llm or OpenAI(model=DEFAULT_MODEL_NAME)
         if callback_manager is not None:
             llm.callback_manager = callback_manager
-
-        memory = memory or memory_cls.from_defaults(chat_history=chat_history, llm=llm)
-
+        memory = memory or memory_cls.from_defaults(
+            chat_history=chat_history or [], llm=llm
+        )
         return cls(
-            tools=tools,
+            tools=tools or [],
             tool_retriever=tool_retriever,
             llm=llm,
             memory=memory,
@@ -106,6 +102,7 @@ class ReActAgent(BaseAgent):
             output_parser=output_parser,
             callback_manager=callback_manager,
             verbose=verbose,
+            **kwargs,
         )
 
     @property
@@ -172,6 +169,8 @@ class ReActAgent(BaseAgent):
             tool_output = tool.call(**reasoning_step.action_input)
             event.on_end(payload={EventPayload.FUNCTION_OUTPUT: str(tool_output)})
 
+        self.sources.append(tool_output)
+
         observation_step = ObservationReasoningStep(observation=str(tool_output))
         current_reasoning.append(observation_step)
         if self._verbose:
@@ -199,6 +198,9 @@ class ReActAgent(BaseAgent):
         ) as event:
             tool_output = await tool.acall(**reasoning_step.action_input)
             event.on_end(payload={EventPayload.FUNCTION_OUTPUT: str(tool_output)})
+
+        self.sources.append(tool_output)
+
         observation_step = ObservationReasoningStep(observation=str(tool_output))
         current_reasoning.append(observation_step)
         if self._verbose:
@@ -218,7 +220,7 @@ class ReActAgent(BaseAgent):
         response_step = cast(ResponseReasoningStep, current_reasoning[-1])
 
         # TODO: add sources from reasoning steps
-        return AgentChatResponse(response=response_step.response, sources=[])
+        return AgentChatResponse(response=response_step.response, sources=self.sources)
 
     @trace_method("chat")
     def chat(
@@ -227,6 +229,7 @@ class ReActAgent(BaseAgent):
         """Chat."""
         # get tools
         # TODO: do get tools dynamically at every iteration of the agent loop
+        self.sources = []
         tools = self.get_tools(message)
 
         if chat_history is not None:
@@ -265,6 +268,7 @@ class ReActAgent(BaseAgent):
     ) -> AgentChatResponse:
         # get tools
         # TODO: do get tools dynamically at every iteration of the agent loop
+        self.sources = []
         tools = self.get_tools(message)
 
         if chat_history is not None:
@@ -303,6 +307,7 @@ class ReActAgent(BaseAgent):
     ) -> StreamingAgentChatResponse:
         # get tools
         # TODO: do get tools dynamically at every iteration of the agent loop
+        self.sources = []
         tools = self.get_tools(message)
 
         if chat_history is not None:
@@ -341,7 +346,9 @@ class ReActAgent(BaseAgent):
             current_reasoning.extend(reasoning_steps)
 
         # Get the response in a separate thread so we can yield the response
-        chat_stream_response = StreamingAgentChatResponse(chat_stream=chat_stream)
+        chat_stream_response = StreamingAgentChatResponse(
+            chat_stream=chat_stream, sources=self.sources
+        )
         thread = Thread(
             target=chat_stream_response.write_response_to_history,
             args=(self._memory,),
@@ -355,6 +362,7 @@ class ReActAgent(BaseAgent):
     ) -> StreamingAgentChatResponse:
         # get tools
         # TODO: do get tools dynamically at every iteration of the agent loop
+        self.sources = []
         tools = self.get_tools(message)
 
         if chat_history is not None:
@@ -394,7 +402,9 @@ class ReActAgent(BaseAgent):
             current_reasoning.extend(reasoning_steps)
 
         # Get the response in a separate thread so we can yield the response
-        chat_stream_response = StreamingAgentChatResponse(achat_stream=chat_stream)
+        chat_stream_response = StreamingAgentChatResponse(
+            achat_stream=chat_stream, sources=self.sources
+        )
         # create task to write chat response to history
         asyncio.create_task(
             chat_stream_response.awrite_response_to_history(self._memory)