+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # ChangeLog
 
+## Unreleased
+
+### Bug Fixes / Nits
+
+- Add normalization to huggingface embeddings (#8145)
+
 ## [0.8.45] - 2023-10-13
 
 ### New Features
@@ -14,6 +20,7 @@
 - Improved the `BM25Retriever` interface to accept `BaseNode` objects (#8096)
 - Fixed bug with `BM25Retriever` tokenizer not working as expected (#8096)
 - Brought mypy to pass in Python 3.8 (#8107)
+- `ReActAgent` adding missing `super().__init__` call (#8125)
 
 ## [0.8.44] - 2023-10-12